move steam backend logic behind platform service
