+++ src-tauri/src/platforms/mod.rs
@@ -1,0 +1,1 @@
+pub mod steam;
